clippy: apply cargo clippy --fix across the AI-PoW crates
Machine-applicable style fixes only (behaviour-preserving): `x % n == 0` ->
`x.is_multiple_of(n)`, `acc = acc + x` -> `acc += x`, unused-binding underscoring,
unused-import removal, and doc-list continuation markers. No logic changes.

+++ crates/ai-pow-zk/src/composite_proof.rs
@@ -410,7 +410,7 @@ pub(crate) fn logup_common_for(
     use p3_batch_stark::ProverData;
     let log_ext_db = checked_program_degree_bits(program)
         .expect("canonical program shape already validated")
-        + config.is_zk() as usize;
+        + config.is_zk();
     let air =
         crate::composite_full_air_with_lookups::CompositeFullAirWithLookupsPinned::try_new_with(
             program.clone(),
